+++ src/test/java/calculator/view/InputValidatorTest.java
@@ -1,6 +1,5 @@
-package calculator.utils;
+package calculator.view;
 
-import calculator.model.InputValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
