+++ src/test/java/racingcargame/view/RaceCountInputValidatorTest.java
@@ -1,10 +1,9 @@
-package racingcargame.utils;
+package racingcargame.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcargame.view.RaceCountInputValidator;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
